refactor: Moved LabeledSlider into a new `Inputs` folder
